Creates a RPC module with serialization; Only AccountInfo is supported at the time
