--- networking/src/commonMain/kotlin/foundation/metaplex/networking/NetworkClient.kt
@@ -1,5 +1,0 @@
-package foundation.metaplex.networking
-
-import io.ktor.client.HttpClient
-
-fun NetworkClient(): HttpClient = HttpClient()
