+++ rpc/src/commonMain/kotlin/foundation/metaplex/rpc/networking/NetworkClient.kt
@@ -1,0 +1,6 @@
+package foundation.metaplex.rpc.networking
+
+import io.ktor.client.HttpClient
+
+// Standard NetworkClient. Used as Default
+fun NetworkClient(): HttpClient = HttpClient()
