--- serializers/src/commonMain/kotlin/foundation/metaplex/serializers/BorshAsBase64JsonArraySerializer.kt
@@ -1,39 +1,0 @@
-package foundation.metaplex.serializers
-
-import com.funkatronics.kborsh.BorshDecoder
-import com.funkatronics.kborsh.BorshEncoder
-import kotlinx.serialization.KSerializer
-import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.encoding.Decoder
-import kotlinx.serialization.encoding.Encoder
-
-/**
- * Decodes/Encodes input using the Borsh encoding scheme, and serializes it as a Base64 encoded
- * string, formatted as Json string array:
- * output = {
- *      [
- *          "theBorshEncodedBytesAsBase64String",
- *          "base64"
- *      ]
- * }
- */
-class BorshAsBase64JsonArraySerializer<T>(private val dataSerializer: KSerializer<T>):
-    KSerializer<T?> {
-    private val delegateSerializer = ByteArrayAsBase64JsonArraySerializer
-    override val descriptor: SerialDescriptor = dataSerializer.descriptor
-
-    override fun serialize(encoder: Encoder, value: T?) =
-        encoder.encodeSerializableValue(delegateSerializer,
-            value?.let {
-                BorshEncoder().apply {
-                    encodeSerializableValue(dataSerializer, value)
-                }.borshEncodedBytes
-            } ?: byteArrayOf()
-        )
-
-    override fun deserialize(decoder: Decoder): T? =
-        decoder.decodeSerializableValue(delegateSerializer).run {
-            if (this.isEmpty()) return null
-            BorshDecoder(this).decodeSerializableValue(dataSerializer)
-        }
-}
