--- serializers/src/commonMain/kotlin/foundation/metaplex/serializers/SolanaResponseSerializer.kt
@@ -1,22 +1,0 @@
-package foundation.metaplex.serializers
-
-import kotlinx.serialization.KSerializer
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.encoding.Decoder
-import kotlinx.serialization.encoding.Encoder
-
-class SolanaResponseSerializer<R>(dataSerializer: KSerializer<R>)
-    : KSerializer<R?> {
-    private val serializer = WrappedValue.serializer(dataSerializer)
-    override val descriptor: SerialDescriptor = serializer.descriptor
-
-    override fun serialize(encoder: Encoder, value: R?) =
-        encoder.encodeSerializableValue(serializer, WrappedValue(value))
-
-    override fun deserialize(decoder: Decoder): R? =
-        decoder.decodeSerializableValue(serializer).value
-}
-
-@Serializable
-private class WrappedValue<V>(val value: V?)
