+++ settings.gradle.kts
@@ -23,5 +23,4 @@ include(":solanainterfaces")
 include(":solanakeypair")
 include(":amount")
 include(":readapi")
-include(":serializers")
-include(":networking")
+include(":rpc")
